+++ packages/nestjs/src/index.ts
@@ -2,6 +2,6 @@ export * from '@sentry/node';
 
 export { init } from './sdk';
 
-export { SentryTraced } from './span-decorator';
-export { SentryCron } from './cron-decorator';
-export { WithSentry } from './error-decorator';
+export { SentryTraced } from './decorators/span-decorator';
+export { SentryCron } from './decorators/cron-decorator';
+export { WithSentry } from './decorators/error-decorator';
